+++ configure.py
@@ -720,6 +720,7 @@ scylla_core = (['message/messaging_service.cc',
                 'sstables/sstables.cc',
                 'sstables/sstables_manager.cc',
                 'sstables/sstable_set.cc',
+                'sstables/storage.cc',
                 'sstables/mx/partition_reversing_data_source.cc',
                 'sstables/mx/reader.cc',
                 'sstables/mx/writer.cc',
